fix `runReplayForPlugin` check after API classes migration
