+++ src/jvm/main/org/jetbrains/lincheck/datastructures/ModelCheckingStrategyConfiguration.kt
@@ -73,7 +73,7 @@ class ModelCheckingCTestConfiguration(
     sequentialSpecification: Class<*>,
     timeoutMs: Long,
     customScenarios: List<ExecutionScenario>,
-    internal val stdLibAnalysisEnabled: Boolean,
+    stdLibAnalysisEnabled: Boolean,
 ) : ManagedCTestConfiguration(
     testClass = testClass,
     iterations = iterations,
@@ -90,7 +90,8 @@ class ModelCheckingCTestConfiguration(
     minimizeFailedScenario = minimizeFailedScenario,
     sequentialSpecification = sequentialSpecification,
     timeoutMs = timeoutMs,
-    customScenarios = customScenarios
+    customScenarios = customScenarios,
+    stdLibAnalysisEnabled = stdLibAnalysisEnabled,
 ) {
 
     override val instrumentationMode: InstrumentationMode get() = MODEL_CHECKING
@@ -107,13 +108,4 @@ class ModelCheckingCTestConfiguration(
         stateRepresentationMethod,
         createSettings()
     )
-
-    internal fun createSettings(): ManagedStrategySettings =
-        ManagedStrategySettings(
-            timeoutMs = this.timeoutMs,
-            hangingDetectionThreshold = this.hangingDetectionThreshold,
-            checkObstructionFreedom = this.checkObstructionFreedom,
-            analyzeStdLib = this.stdLibAnalysisEnabled,
-            guarantees = this.guarantees.ifEmpty { null },
-        )
 }
